Move parseParams
