+++ lib/elasticsearch/config.js
@@ -129,8 +129,47 @@ const AGGREGATIONS_SPEC = {
   collection: { terms: { field: 'collectionIds' } }
 }
 
+const ITEM_FILTER_AGGREGATIONS = {
+  item_location: { nested: { path: 'items' }, aggs: { _nested: { terms: { size: 100, field: 'items.holdingLocation_packed' } } } },
+  item_status: { nested: { path: 'items' }, aggs: { _nested: { terms: { size: 100, field: 'items.status_packed' } } } },
+  item_format: { nested: { path: 'items' }, aggs: { _nested: { terms: { size: 100, field: 'items.formatLiteral' } } } }
+}
+
+// Configure sort fields:
+const SORT_FIELDS = {
+  title: {
+    initialDirection: 'asc',
+    field: 'title_sort'
+  },
+  date: {
+    initialDirection: 'desc',
+    field: 'dateStartYear'
+  },
+  creator: {
+    initialDirection: 'asc',
+    field: 'creator_sort'
+  },
+  relevance: {}
+}
+
+// The following fields can be excluded from ES responses because we don't pass them to client:
+const EXCLUDE_FIELDS = [
+  'uris',
+  '*_packed',
+  '*_sort',
+  'items.*_packed',
+  'contentsTitle',
+  'suppressed',
+  // Hide contributor and creator transformed fields:
+  '*WithoutDates',
+  '*Normalized'
+]
+
 module.exports = {
   SEARCH_SCOPES,
   FILTER_CONFIG,
-  AGGREGATIONS_SPEC
+  AGGREGATIONS_SPEC,
+  ITEM_FILTER_AGGREGATIONS,
+  EXCLUDE_FIELDS,
+  SORT_FIELDS
 }
