+++ lib/resources.js
@@ -19,77 +19,12 @@ const { parseParams, deepValue } = require('../lib/util')
 
 const ApiRequest = require('./api-request')
 const ElasticQueryBuilder = require('./elasticsearch/elastic-query-builder')
-const { FILTER_CONFIG, SEARCH_SCOPES, AGGREGATIONS_SPEC } = require('./elasticsearch/config')
+const { AGGREGATIONS_SPEC, ITEM_FILTER_AGGREGATIONS, EXCLUDE_FIELDS, SORT_FIELDS } = require('./elasticsearch/config')
 
 const errors = require('./errors')
-const { esRangeValue } = require('./utils/resource-helpers')
+const { esRangeValue, parseSearchParams } = require('./utils/resource-helpers')
 
 const RESOURCES_INDEX = process.env.RESOURCES_INDEX
-
-const ITEM_FILTER_AGGREGATIONS = {
-  item_location: { nested: { path: 'items' }, aggs: { _nested: { terms: { size: 100, field: 'items.holdingLocation_packed' } } } },
-  item_status: { nested: { path: 'items' }, aggs: { _nested: { terms: { size: 100, field: 'items.status_packed' } } } },
-  item_format: { nested: { path: 'items' }, aggs: { _nested: { terms: { size: 100, field: 'items.formatLiteral' } } } }
-}
-
-// Configure sort fields:
-const SORT_FIELDS = {
-  title: {
-    initialDirection: 'asc',
-    field: 'title_sort'
-  },
-  date: {
-    initialDirection: 'desc',
-    field: 'dateStartYear'
-  },
-  creator: {
-    initialDirection: 'asc',
-    field: 'creator_sort'
-  },
-  relevance: {}
-}
-
-// The following fields can be excluded from ES responses because we don't pass them to client:
-const EXCLUDE_FIELDS = [
-  'uris',
-  '*_packed',
-  '*_sort',
-  'items.*_packed',
-  'contentsTitle',
-  'suppressed',
-  // Hide contributor and creator transformed fields:
-  '*WithoutDates',
-  '*Normalized'
-]
-
-// Configure controller-wide parameter parsing:
-const parseSearchParams = function (params, overrideParams = {}) {
-  return parseParams(params, {
-    q: { type: 'string' },
-    page: { type: 'int', default: 1 },
-    per_page: { type: 'int', default: 50, range: [0, 100] },
-    field: { type: 'string', range: Object.keys(AGGREGATIONS_SPEC) },
-    sort: { type: 'string', range: Object.keys(SORT_FIELDS), default: 'relevance' },
-    sort_direction: { type: 'string', range: ['asc', 'desc'] },
-    search_scope: { type: 'string', range: Object.keys(SEARCH_SCOPES), default: 'all' },
-    filters: { type: 'hash', fields: FILTER_CONFIG },
-    items_size: { type: 'int', default: 100, range: [0, 200] },
-    items_from: { type: 'int', default: 0 },
-    callnumber: { type: 'string' },
-    standard_number: { type: 'string' },
-    contributor: { type: 'string' },
-    title: { type: 'string' },
-    subject: { type: 'string' },
-    subject_prefix: { type: 'string' },
-    isbn: { type: 'string' },
-    issn: { type: 'string' },
-    lccn: { type: 'string' },
-    oclc: { type: 'string' },
-    merge_checkin_card_items: { type: 'boolean', default: true },
-    include_item_aggregations: { type: 'boolean', default: true },
-    ...overrideParams
-  })
-}
 
 // These are the handlers made available to the router:
 module.exports = function (app, _private = null) {
