+++ lib/utils/resource-helpers.js
@@ -1,3 +1,6 @@
+const { parseParams } = require('../util')
+const { FILTER_CONFIG, SEARCH_SCOPES, AGGREGATIONS_SPEC, SORT_FIELDS } = require('../elasticsearch/config')
+
 /**
  * Given a range represented as an array, returns a corresponding ES range object
  *
@@ -24,6 +27,36 @@ const esRangeValue = (range) => {
   return rangeQuery
 }
 
+// Configure controller-wide parameter parsing:
+const parseSearchParams = function (params, overrideParams = {}) {
+  return parseParams(params, {
+    q: { type: 'string' },
+    page: { type: 'int', default: 1 },
+    per_page: { type: 'int', default: 50, range: [0, 100] },
+    field: { type: 'string', range: Object.keys(AGGREGATIONS_SPEC) },
+    sort: { type: 'string', range: Object.keys(SORT_FIELDS), default: 'relevance' },
+    sort_direction: { type: 'string', range: ['asc', 'desc'] },
+    search_scope: { type: 'string', range: Object.keys(SEARCH_SCOPES), default: 'all' },
+    filters: { type: 'hash', fields: FILTER_CONFIG },
+    items_size: { type: 'int', default: 100, range: [0, 200] },
+    items_from: { type: 'int', default: 0 },
+    callnumber: { type: 'string' },
+    standard_number: { type: 'string' },
+    contributor: { type: 'string' },
+    title: { type: 'string' },
+    subject: { type: 'string' },
+    subject_prefix: { type: 'string' },
+    isbn: { type: 'string' },
+    issn: { type: 'string' },
+    lccn: { type: 'string' },
+    oclc: { type: 'string' },
+    merge_checkin_card_items: { type: 'boolean', default: true },
+    include_item_aggregations: { type: 'boolean', default: true },
+    ...overrideParams
+  })
+}
+
 module.exports = {
-  esRangeValue
+  esRangeValue,
+  parseSearchParams
 }
